v8.0.4

+++ src/lib/stores/player.ts
@@ -43,7 +43,7 @@ const createInitialState = (): PlayerStore => ({
   fullDuration: 0,
   playbackRate: 1.0,
   loop: false,
-  volume: parseInt(config.volume) / 100,
+  volume: parseFloat(config.volume) / 100,
   stream: {
     title: '',
     author: '',
